implemented login connection to the backend
